Fix type definition error in PlanetEntity.m
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5568 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/PlanetEntity.h
@@ -110,7 +110,7 @@ typedef struct
 - (void) setPlanetType:(OOStellarBodyType) pt;
 
 
-- (GLfloat) radius;	// metres
+- (double) radius;	// metres
 - (void) setRadius:(GLfloat) rad;
 - (double) rotationalVelocity;
 - (void) setRotationalVelocity:(double) v;
